+++ packages/frontend/src/components/chat/ChatListItem.tsx
@@ -108,7 +108,9 @@ const Message = React.memo<
           <div
             className='summary_thumbnail'
             style={{
-              backgroundImage: `url(${JSON.stringify(summaryPreviewImage)})`,
+              backgroundImage: `url("${JSON.stringify(
+                runtime.transformBlobURL(summaryPreviewImage)
+              )}")`,
             }}
           />
         )}