+++ packages/shared/shared-types.d.ts
@@ -80,6 +80,8 @@ export type RuntimeInfo = {
   isMac: boolean
   /** currently used to check for an additional device message */
   isAppx: boolean
+  /** to show / hide elements/options that are not supported, like tray icon options on browser */
+  target: 'electron' | 'browser' | 'tauri'
   /** runtime library versions, be it electron, node, tauri or whatever,
    *  used for showing to user in the About dialog */
   versions: { label: string; value: string }[]